+++ CG_GERATOR/FactoryTunnel.cpp
@@ -1,6 +1,7 @@
+#include "stdafx.h"
 #include "FigureFactory.h"
 
-Figure FigureFactory::createTunnel(float insideRadius, float outsideRadius, float nsides, float rings){
+Figure FigureFactory::createTunnel(float insideRadius, float outsideRadius, int nsides, int rings){
 	Figure f;
 	return f;
 }
